Разделила на модули и прописала для них Докстринги
